% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jihong.R
\name{jihong}
\alias{jihong}
\title{This is the function for Jihong Zhang}
\usage{
jihong(details = FALSE)
}
\arguments{
\item{details}{Want to know more}
}
\value{
descripe some information about Jihong Zhang
}
\description{
This is the function for Jihong Zhang
}
\examples{
jihong(details = TRUE)
}
